ETL/Extract/requirements.txt
# Requirements for extract_bulk_trip_record_data.py & extract_incremental_load_trip_record_data.py
requests==2.32.3
beautifulsoup4==4.12.3
google-cloud-storage==2.19.0

#Requirements for extract_greenhouse_gas_emissions.py & extract_air_quality.py

sodapy==2.2.0
google-cloud-storage==2.19.0

#Requirements for extract_incremental_load_weather
openmeteo_requests==1.3.0
requests_cache==1.2.1
retry_requests==2.0.0
pandas==2.2.3
numpy==2.0.2
google-cloud-storage==2.19.0
